remove depricated navbar
