+++ packages/ui/src/components/primitives/Navbar.stories.tsx
@@ -1,0 +1,16 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import { NavbarRenderer } from './Navbar';
+const meta = {
+	component: NavbarRenderer,
+} as Meta<typeof NavbarRenderer>;
+
+export default meta;
+
+type Story = StoryObj<typeof meta>;
+
+export const NotSignedIn: Story = {
+	args: {
+		path: '/',
+	},
+};
